+++ ttp_templates/__init__.py
@@ -1,3 +1,4 @@
 from .ttp_templates import get_template
 from .ttp_templates import parse_output
 from .ttp_templates import list_templates
+from .ttp_templates import list_templates_refs